Add TRB ring structure
The DbC driver has to manage three TRB rings: one event and two transfer.
Each ring has a enqueue and dequeue offset value into the TRB buffer
as well as a PCS and CCS fields for interpreting the cycle bit correctly.

The struct trb_ring definition assumes only one segment, which is
defined in the spec as a contiguous block of TRBs. Technically a ring
can contain multiple, discontiguous segments, but we stick with one for
simplicity. Each segment can be up to 64KB, but we start with a limit
of 4KB.