Migrate to header-only xue.h
This starts the migration to a single xue.h library. All
types, functions, and #defines are prefixed with 'xue_' or 'XUE_'
to prevent future name collisions.

To use the library, users must define one instance of struct xue_ops and
struct xue each and pass them both to xue_init.