Buffer data from xue_write
If users call xue_write, then copy the data into the work ring.
If the hardware is ready for a transfer, send the current item
in the work ring. If not, then do what you can do get the hw ready
(e.g. clearing DRC bit) and return the number of bytes copied to
the work ring.

This behavior allows the debug target to write to the device prior
to successful enumeration by the debug host. Once configured, the
appropriate transfer TRBs are scheduled on the OUT ring.