Add read.sh script
The debug host must configure the debug target before reading its
output. This script polls the USB until the device file provided
as the first argument exists. Once it does, it cats it to stdout.

So if the debug device will appear at /dev/ttyUSB1, then the usage
would be

    ./read.sh /dev/ttyUSB1