Add Linux and Windows cases in read.py
On Linux, read.py ensures that the DbC was bound to usb_debug then
cat's the tty that usb_debug creates.

On Windows, set_configuration() is called after the DbC is found, then
read() is called repeatedly. This assumes that WinUSB is bound to the
DbC so that the libusb backend can do IO with it.