Memory management for xl create
This change adds support of the various XENMEM_* hypercalls that
xl performs as it builds a new domain. Note that only "auto-translated"
domains are supported, which is xen-speak for hardware-assisted paging
or EPT. Most of the XENMEM hypercalls made modify the domains EPT tables
in some way.

To support these calls, a new xen_page class is created that abstracts
an actual physical page that has been assigned to a domain. Each
xen_page can have a backing page. Currently there are two sources of
backing pages: the root domain and the VMM. Pages from the root domain
are used first, and if none are available then pages are allocated from
the VMMs page pool.