deps/xue: Fix link TRB bug
Ensure that the xHC acquires ownership of the tranfer ring's link
TRB by setting the cycle bit to the value of ring->cyc (prior to
toggline ring->cyc). This ensures the wrap-around of the transfer
ring works properly.

Also add bytes_written and trbs_written members to struct xue
for debugging purposes and adjust the xue_printf macro definition
for debugging from the microv vmm. The new version of xue_printf
writes to the vmm's debug ring rather than the COM port so that
'make dump' can be used to view xue state.